assets/module4/back/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via cloud-sql-python-connector
aiosignal==1.3.1
    # via aiohttp
asn1crypto==1.5.1
    # via scramp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.6.2
    # via flask
cachetools==5.3.0
    # via google-auth
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via flask
cloud-sql-python-connector[pg8000]==1.2.3
    # via -r requirements.in
cryptography==40.0.2
    # via cloud-sql-python-connector
flask==2.3.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
google-auth==2.18.0
    # via cloud-sql-python-connector
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pg8000==1.29.4
    # via cloud-sql-python-connector
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
python-dateutil==2.8.2
    # via pg8000
requests==2.30.0
    # via cloud-sql-python-connector
rsa==4.9
    # via google-auth
scramp==1.4.4
    # via pg8000
six==1.16.0
    # via
    #   google-auth
    #   python-dateutil
sqlalchemy==2.0.13
    # via -r requirements.in
typing-extensions==4.5.0
    # via sqlalchemy
urllib3==1.26.15
    # via
    #   google-auth
    #   requests
werkzeug==2.3.4
    # via flask
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
